@@ -19,77 +18,41 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 version: 1.0.0+1
 
 environment:
-  sdk: '>=3.1.5 <3.24.3'
+  sdk: '>=3.3.3 <4.0.0'
 
-# Dependencies specify other packages that your package needs in order to work.
-# To automatically upgrade your package dependencies to the latest versions
-# consider running `flutter pub upgrade --major-versions`. Alternatively,
-# dependencies can be manually updated by changing the version numbers below to
-# the latest version available on pub.dev. To see which dependencies have newer
-# versions available, run `flutter pub outdated`.
 dependencies:
   flutter:
     sdk: flutter
 
-  # The following adds the Cupertino Icons font to your application.
-  # Use with the CupertinoIcons class for iOS style icons.
-  cupertino_icons: ^1.0.2
-
-  # Support for various charts
-  fl_chart: ^0.64.0
-
-  # State manager
-  flutter_riverpod: ^2.1.3
-
-  # Database storage
-  sqflite: ^2.2.2
-  sqflite_common_ffi: ^2.2.0+1
-
-  # iOS, Android and MacOS
+  collection: ^1.18.0
+  cupertino_icons: ^1.0.8
+  fl_chart: ^0.69.0
+  flutter_riverpod: ^2.6.1
+  flutter_native_splash: ^2.4.2
+  flutter_local_notifications: ^18.0.0
+  intl: ^0.19.0
+  path: ^1.9.0
+  percent_indicator: ^4.2.3
+  permission_handler: ^11.3.1
+  shared_preferences: ^2.3.3
+  sqflite: ^2.4.1
+  sqflite_common_ffi: ^2.3.4
   sqlite3_flutter_libs:
-
-  # mailto
-  url_launcher: ^6.0.6
-
-  # Format DateTime
-  intl: ^0.18.1
-
-  # Display todos in grid
-  flutter_staggered_grid_view: ^0.7.0
-
-  percent_indicator: ^4.2.2
-  circular_menu: ^2.0.1
-  flutter_native_splash: ^2.2.18
-  shared_preferences: ^2.2.2
-  flutter_local_notifications: 17.2.3
-  permission_handler: 11.2.0
-  workmanager: 0.5.2
+  timezone: ^0.9.4
+  url_launcher: ^6.3.1
+  workmanager: ^0.5.2
 
 dev_dependencies:
   flutter_test:
     sdk: flutter
 
-  flutter_launcher_icons: ^0.13.1
+  dependency_validator: ^4.1.1
+  flutter_launcher_icons: ^0.14.1
+  flutter_lints: ^5.0.0
+  test: ^1.25.2
 
-  # The "flutter_lints" package below contains a set of recommended lints to
-  # encourage good coding practices. The lint set provided by the package is
-  # activated in the `analysis_options.yaml` file located at the root of your
-  # package. See that file for information about deactivating specific lint
-  # rules and activating additional ones.
-  flutter_lints: ^3.0.1
-
-  # Flutter test
-  test: ^1.5.1
-
-# For information on the generic Dart part of this file, see the
-# following page: https://dart.dev/tools/pub/pubspec
-
-# The following section is specific to Flutter packages.
 flutter:
 
-  # The following line ensures that the Material Icons font is
-  # included with your application, so that you can use the icons in
-  # the material Icons class.
   uses-material-design: true
 
   fonts:
