@@ -1,5 +1,3 @@
-import 'dart:ui';
-
 import 'package:fl_chart/fl_chart.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
@@ -17,7 +15,7 @@ class AccountDialog extends ConsumerWidget with Functions {
   @override
   Widget build(BuildContext context, WidgetRef ref) {
     final currencyState = ref.watch(currencyStateNotifier);
-    
+
     return SingleChildScrollView(
       child: Column(
         children: [
@@ -62,7 +60,7 @@ class AccountDialog extends ConsumerWidget with Functions {
                   padding: EdgeInsets.all(8.0),
                 ),
                 LineChartWidget(
-                  lineData: [
+                  lineData: const [
                     FlSpot(0, 3),
                     FlSpot(1, 1.3),
                     FlSpot(2, -2),
@@ -81,15 +79,8 @@ class AccountDialog extends ConsumerWidget with Functions {
                     FlSpot(15, -4.5),
                     FlSpot(16, 2.5),
                   ],
-                  lineColor: Color(0xffffffff),
-                  line2Data: <FlSpot>[],
-                  line2Color: Color(0xffffffff),
-                  colorBackground: Color(0xff356CA3),
-                  period: Period.month,
                 ),
-                const Padding(
-                  padding: EdgeInsets.only(bottom: 14.0),
-                ),
+                const SizedBox(height: 14.0),
               ],
             ),
           ),
