@@ -52,7 +52,7 @@ class _AccountListState extends ConsumerState<AccountList> with Functions {
                     child: Icon(
                       Icons.account_balance_wallet,
                       size: 24.0,
-                      color: Theme.of(context).colorScheme.background,
+                      color: Theme.of(context).colorScheme.onPrimary,
                     ),
                   ),
                   const SizedBox(width: 12.0),
@@ -95,7 +95,7 @@ class _AccountListState extends ConsumerState<AccountList> with Functions {
                               ? Icon(
                                   icon,
                                   size: 30.0,
-                                  color: Theme.of(context).colorScheme.background,
+                                  color: Theme.of(context).colorScheme.onPrimary,
                                 )
                               : const SizedBox(),
                         ),
