@@ -1,6 +1,5 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
-import 'package:intl/intl.dart';
 
 import '../../constants/style.dart';
 import '../../model/currency.dart';
@@ -12,8 +11,7 @@ class GeneralSettingsPage extends ConsumerStatefulWidget {
   const GeneralSettingsPage({super.key});
 
   @override
-  ConsumerState<GeneralSettingsPage> createState() =>
-      _GeneralSettingsPageState();
+  ConsumerState<GeneralSettingsPage> createState() => _GeneralSettingsPageState();
 }
 
 class _GeneralSettingsPageState extends ConsumerState<GeneralSettingsPage> {
@@ -58,8 +56,10 @@ class _GeneralSettingsPageState extends ConsumerState<GeneralSettingsPage> {
             Row(
               children: [
                 Text("Appearance",
-                    style: Theme.of(context).textTheme.titleLarge!.copyWith(
-                        color: Theme.of(context).colorScheme.primary)),
+                    style: Theme.of(context)
+                        .textTheme
+                        .titleLarge!
+                        .copyWith(color: Theme.of(context).colorScheme.primary)),
                 const Spacer(),
                 CircleAvatar(
                     radius: 30.0,
@@ -69,21 +69,15 @@ class _GeneralSettingsPageState extends ConsumerState<GeneralSettingsPage> {
                       onPressed: () {
                         // Toggle dark mode using the provider
                         if (appThemeState.isDarkModeEnabled) {
-                          ref
-                              .read(appThemeStateNotifier.notifier)
-                              .setLightTheme();
+                          ref.read(appThemeStateNotifier.notifier).setLightTheme();
                         } else {
-                          ref
-                              .read(appThemeStateNotifier.notifier)
-                              .setDarkTheme();
+                          ref.read(appThemeStateNotifier.notifier).setDarkTheme();
                         }
                       },
                       icon: Icon(
-                        appThemeState.isDarkModeEnabled
-                            ? Icons.dark_mode
-                            : Icons.light_mode,
+                        appThemeState.isDarkModeEnabled ? Icons.dark_mode : Icons.light_mode,
                         size: 25.0,
-                        color: Theme.of(context).colorScheme.background,
+                        color: Theme.of(context).colorScheme.onPrimary,
                       ),
                     )),
               ],
@@ -92,13 +86,16 @@ class _GeneralSettingsPageState extends ConsumerState<GeneralSettingsPage> {
             Row(
               children: [
                 Text("Currency",
-                    style: Theme.of(context).textTheme.titleLarge!.copyWith(
-                        color: Theme.of(context).colorScheme.primary)),
+                    style: Theme.of(context)
+                        .textTheme
+                        .titleLarge!
+                        .copyWith(color: Theme.of(context).colorScheme.primary)),
                 const Spacer(),
                 GestureDetector(
                     onTap: () {
                       setState(() {
-                        CurrencySelectorDialog.selectCurrencyDialog(context, currencyState, currencyList);
+                        CurrencySelectorDialog.selectCurrencyDialog(
+                            context, currencyState, currencyList);
                       });
                     },
                     child: CircleAvatar(
@@ -108,8 +105,7 @@ class _GeneralSettingsPageState extends ConsumerState<GeneralSettingsPage> {
                             child: Text(
                           currencyState.selectedCurrency.symbol,
                           style: TextStyle(
-                              color: Theme.of(context).colorScheme.background,
-                              fontSize: 25),
+                              color: Theme.of(context).colorScheme.onPrimary, fontSize: 25),
                         )))),
               ],
             ),
