@@ -75,12 +75,13 @@ var collaborators = const [
   ]
 ];
 
+
 class _CollaboratorsPageState extends ConsumerState<CollaboratorsPage> {
   @override
   Widget build(BuildContext context) {
     return Scaffold(
       appBar: AppBar(
-        backgroundColor: Theme.of(context).colorScheme.background,
+        backgroundColor: Theme.of(context).colorScheme.onPrimary,
         elevation: 0,
         centerTitle: true,
         leading: IconButton(
@@ -102,16 +103,16 @@ class _CollaboratorsPageState extends ConsumerState<CollaboratorsPage> {
         ),
       ),
       body: SingleChildScrollView(
-          physics: const BouncingScrollPhysics(),
-          child: Padding(
-            padding: EdgeInsets.all(8.0),
+        physics: const BouncingScrollPhysics(),
+        child:
+          Padding(padding: const EdgeInsets.all(8.0),
             child: Column(
               children: [
                 ListView.separated(
                   itemCount: collaborators.length,
                   shrinkWrap: true,
                   physics: const NeverScrollableScrollPhysics(),
-                  separatorBuilder: (context, index) => Divider(),
+                  separatorBuilder: (context, index) => const Divider(),
                   itemBuilder: (context, i) {
                     List option = collaborators[i];
                     return InkWell(
@@ -126,37 +127,19 @@ class _CollaboratorsPageState extends ConsumerState<CollaboratorsPage> {
                           children: [
                             Text(
                               option[0].toString(),
-                              style: Theme.of(context)
-                                  .textTheme
-                                  .titleLarge!
-                                  .copyWith(
-                                      color: Theme.of(context)
-                                          .colorScheme
-                                          .primary),
+                              style: Theme.of(context).textTheme.titleLarge!.copyWith(color: Theme.of(context).colorScheme.primary),
                               textAlign: TextAlign.left,
                             ),
                             const SizedBox(height: 4),
                             Text(
                               option[1].toString(),
-                              style: Theme.of(context)
-                                  .textTheme
-                                  .bodyMedium!
-                                  .copyWith(
-                                      color: Theme.of(context)
-                                          .colorScheme
-                                          .primary),
+                              style: Theme.of(context).textTheme.bodyMedium!.copyWith(color: Theme.of(context).colorScheme.primary),
                               textAlign: TextAlign.left,
                             ),
                             const SizedBox(height: 4),
                             Text(
                               option[2].toString(),
-                              style: Theme.of(context)
-                                  .textTheme
-                                  .bodySmall!
-                                  .copyWith(
-                                      color: Theme.of(context)
-                                          .colorScheme
-                                          .primary),
+                              style: Theme.of(context).textTheme.bodySmall!.copyWith(color: Theme.of(context).colorScheme.primary),
                               textAlign: TextAlign.left,
                             ),
                           ],
@@ -165,9 +148,11 @@ class _CollaboratorsPageState extends ConsumerState<CollaboratorsPage> {
                     );
                   },
                 ),
+
               ],
             ),
-          )),
+          )
+      ),
     );
   }
 }
