@@ -1,5 +1,3 @@
-import 'dart:io';
-
 import 'package:flutter_test/flutter_test.dart';
 import 'package:intl/intl.dart';
 import 'package:sossoldi/database/sossoldi_database.dart';
@@ -83,7 +81,7 @@ void main() {
     assert((b.mainAccount ? 1 : 0) == json[BankAccountFields.mainAccount]);
   });
 
-  group("Bank Account Methods", () { 
+  group("Bank Account Methods", () {
 
     late SossoldiDatabase sossoldiDatabase;
     late sqflite.Database db;
@@ -127,7 +125,7 @@ void main() {
       final today = DateTime.now();
       final fistOfCurrentMonth = DateTime(today.year, today.month, 1);
 
-      // Add a transaction of last month      
+      // Add a transaction of last month
       demoTransactions.add(createInsertSqlTransaction(date: fistOfCurrentMonth.subtract(const Duration(days: 10))));
       demoTransactions.add(createInsertSqlTransaction(date: fistOfCurrentMonth.subtract(const Duration(days: 10)), idBankAccount: 71));
       demoTransactions.add(createInsertSqlTransaction(date: fistOfCurrentMonth.subtract(const Duration(days: 10)), idBankAccount: 71, type: 'TRSF', idBankTransfert: 70));
@@ -140,7 +138,7 @@ void main() {
       // 2
       demoTransactions.add(createInsertSqlTransaction(date: fistOfCurrentMonth.add(const Duration(days: 1))));
       demoTransactions.add(createInsertSqlTransaction(date: fistOfCurrentMonth.add(const Duration(days: 1)), amount: 200, type: 'IN'));
-      demoTransactions.add(createInsertSqlTransaction(date: fistOfCurrentMonth.add(const Duration(days: 1)), idBankAccount: 71));      
+      demoTransactions.add(createInsertSqlTransaction(date: fistOfCurrentMonth.add(const Duration(days: 1)), idBankAccount: 71));
       demoTransactions.add(createInsertSqlTransaction(date: fistOfCurrentMonth.add(const Duration(days: 1)), amount: 50.5, idBankAccount: 70, type: 'TRSF', idBankTransfert: 71));
       // 3
       demoTransactions.add(createInsertSqlTransaction(date: fistOfCurrentMonth.add(const Duration(days: 2)), type: 'IN'));
@@ -149,7 +147,7 @@ void main() {
 
       // Add recurring transactions. These must be count as number of time they occout * amount
 
-      
+
       await db.execute("$insertDemoTransactionsQuery ${demoTransactions.join(",")};");
 
       transactions = await db.rawQuery("SELECT * FROM `transaction`");
@@ -159,9 +157,9 @@ void main() {
       expect(result.length, 3);
 
       var initialAccountAmount = 1235.10; // taken from fillDemoData
-      expect(result[0].id, 70);   
+      expect(result[0].id, 70);
       expect(result[0].total! - initialAccountAmount, 49.5);
-      
+
       initialAccountAmount = 3823.56; // taken from fillDemoData
       expect(result[1].id, 71);
       expect(result[1].total! - initialAccountAmount, -449.5);
@@ -193,7 +191,7 @@ void main() {
       final today = DateTime.now();
       final fistOfCurrentMonth = DateTime(today.year, today.month, 1);
 
-      // Add a transaction of last month      
+      // Add a transaction of last month
       demoTransactions.add(createInsertSqlTransaction(date: fistOfCurrentMonth.subtract(const Duration(days: 10))));
       demoTransactions.add(createInsertSqlTransaction(date: fistOfCurrentMonth.subtract(const Duration(days: 10)), idBankAccount: 71));
       demoTransactions.add(createInsertSqlTransaction(date: fistOfCurrentMonth.subtract(const Duration(days: 10)), idBankAccount: 71, type: 'TRSF', idBankTransfert: 70));
@@ -201,19 +199,19 @@ void main() {
       // Add transactions of current month
       demoTransactions.add(createInsertSqlTransaction(date: fistOfCurrentMonth));
       demoTransactions.add(createInsertSqlTransaction(date: fistOfCurrentMonth));
-      demoTransactions.add(createInsertSqlTransaction(date: fistOfCurrentMonth, idBankAccount: 71));    
+      demoTransactions.add(createInsertSqlTransaction(date: fistOfCurrentMonth, idBankAccount: 71));
       demoTransactions.add(createInsertSqlTransaction(date: fistOfCurrentMonth.add(const Duration(days: 1))));
       demoTransactions.add(createInsertSqlTransaction(date: fistOfCurrentMonth.add(const Duration(days: 1)), amount: 200, type: 'IN'));
-      demoTransactions.add(createInsertSqlTransaction(date: fistOfCurrentMonth.add(const Duration(days: 1)), idBankAccount: 71));      
+      demoTransactions.add(createInsertSqlTransaction(date: fistOfCurrentMonth.add(const Duration(days: 1)), idBankAccount: 71));
       demoTransactions.add(createInsertSqlTransaction(date: fistOfCurrentMonth.add(const Duration(days: 1)), amount: 50.5, idBankAccount: 70, type: 'TRSF', idBankTransfert: 71));
       demoTransactions.add(createInsertSqlTransaction(date: fistOfCurrentMonth.add(const Duration(days: 2)), type: 'IN'));
       demoTransactions.add(createInsertSqlTransaction(date: fistOfCurrentMonth.add(const Duration(days: 2)), type: 'IN'));
       demoTransactions.add(createInsertSqlTransaction(date: fistOfCurrentMonth.add(const Duration(days: 2)), idBankAccount: 71));
 
-      // Add a transaction of next month      
+      // Add a transaction of next month
       demoTransactions.add(createInsertSqlTransaction(date: fistOfCurrentMonth.add(const Duration(days: 32))));
       demoTransactions.add(createInsertSqlTransaction(date: fistOfCurrentMonth.add(const Duration(days: 32)), idBankAccount: 71));
-      
+
       await db.execute("$insertDemoTransactionsQuery ${demoTransactions.join(",")};");
 
       transactions = await db.rawQuery("SELECT * FROM `transaction`");
@@ -227,7 +225,7 @@ void main() {
 
       final DateFormat formatter = DateFormat('yyyy-MM-dd');
       var initialAccountAmount = 1235.10; // taken from fillDemoData
-      
+
       expect(result[0]['day'], formatter.format(fistOfCurrentMonth));
       expect(result[0]['balance'] - initialAccountAmount, -200);
       expect(result[1]['day'], formatter.format(fistOfCurrentMonth.add(const Duration(days: 1))));
@@ -240,9 +238,9 @@ void main() {
         dateRangeStart: DateTime(DateTime.now().year, DateTime.now().month, 1), // beginnig of current month
         dateRangeEnd: DateTime(DateTime.now().year, DateTime.now().month + 1, 1)); // beginnig of next month;
       expect(result.length, 3);
-      
+
       initialAccountAmount = 3823.56; // taken from fillDemoData
-      
+
       expect(result[0]['day'], formatter.format(fistOfCurrentMonth));
       expect(result[0]['balance'] - initialAccountAmount, -300);
       expect(result[1]['day'], formatter.format(fistOfCurrentMonth.add(const Duration(days: 1))));
@@ -252,6 +250,6 @@ void main() {
     });
   });
 
-  
+
 
 }
